% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linkage_Disequlibrium_Functions.R
\name{calculate_LD}
\alias{calculate_LD}
\title{LD Generation Function}
\usage{
calculate_LD(snp_set_1, snp_set_2, choice, population)
}
\arguments{
\item{snp_set_1}{set of variant 1}

\item{snp_set_2}{set of variant 2}

\item{choice}{"rsids" or "variant_matrix"}

\item{population}{Populations like "EUR","SAS","AFR","AMR","EAS","ALL"}
}
\value{
Linkage
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
snp_set_1 = c("rs112155239","rs62636508","rs58108140","rs574746232","rs571093408")
snp_set_2 = c("rs112155239","rs62636508","rs58108140","rs574746232","rs571093408")
population = "EUR"
choice = "rsids"
calculate_LD(snp_set_1 = snp_set_1 , snp_set_2 = snp_set_2, choice="rsids" ,population = "EUR")


snp_set_1 = c("1_13011","1_10440","1_10583","1_13011")
snp_set_2 = c("1_13011","1_10440","1_10583","1_13011")
population = "EUR"
choice = "variant_matrix"
calculate_LD(snp_set_1 = snp_set_1 , snp_set_2 = snp_set_2, choice="variant_matrix" ,population = "EUR")
}
