% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{g_free}
\alias{g_free}
\title{Closing Function}
\usage{
g_free(con)
}
\arguments{
\item{con}{The connection to the server.}
}
\value{
NA
}
\description{
This Function should be called only once at the end of the program
}
\examples{
g_free(connec)
}
