% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{g_initialize}
\alias{g_initialize}
\title{Initialization Function}
\usage{
g_initialize(func_name)
}
\arguments{
\item{pkg_name}{The name of the packages}
}
\value{
NA
}
\description{
This Function should be called only once before using any function of the package.
}
\examples{
g_initialize("Gtex_Summary");
g_initialize( c("Gtex_Summary", "1000GP_LD_Summary", "Haploreg"));
}
