% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx_Functions.R
\name{Gtex_variance_finder}
\alias{Gtex_variance_finder}
\title{Gtex_variance_finder Function}
\usage{
Gtex_variance_finder(Gene, Tissue)
}
\arguments{
\item{Gene}{Gene Of Interest in the format ENSG00000182179.12}

\item{Tissue}{The Tissue of Interest as per GTEx Portal}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
GTEx_Summary(SNP = "chr3_40_G_T_b38", Gene="ENSG00000182179.12", Tissue="Whole_Blood");
}
