% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_Fucntions.R
\name{Lifthg19Tohg38}
\alias{Lifthg19Tohg38}
\title{Lifthg19Tohg38 Function}
\usage{
Lifthg19Tohg38(Chromosome, Position)
}
\arguments{
\item{rsids}{A vector of Chromosome and a vector of Position beloning to hg19 reference Frame.}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
Lift-over genomic coordinates from hg19 to hg38 position
}
\examples{
Chromosome <- c(8,8,8,8,8,8,8,8,8,8)
Position <- c(101592213,106973048,108690829,108681675,103044620,104152280,100479917,103144592,108166508,105201080)
Lifthg19Tohg38(Chromosome , Position)
}
