% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_Fucntions.R
\name{SNP_To_Allele_1KGP}
\alias{SNP_To_Allele_1KGP}
\title{SNP_To_Allele_1KGP Function}
\usage{
SNP_To_Allele_1KGP(variant_matrix)
}
\arguments{
\item{rsids}{A vector of rsids}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
variant_matrix_2 <- data.frame(Chromosome=c("chr1","chr1","chr1","chr1","chr1"),Position=c(10399,10420,10437,10438,10440))
SNP_To_Allele_1KGP(variant_matrix = variant_matrix_2)
}
