% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haploreg_Functions.R
\name{snpToGENCODE}
\alias{snpToGENCODE}
\title{snpToGENCODE Function}
\usage{
snpToGENCODE(rsids)
}
\arguments{
\item{rsids}{A vector of rsids}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
snpToGENCODE(rsids <- "rs10");
snpToGENCODE(rsids <- c("rs10","rs3","rs4","rs5"));
}
