% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haploreg_Functions.R
\name{snpToInfo}
\alias{snpToInfo}
\title{snpToInfo Function}
\usage{
snpToInfo(rsids)
}
\arguments{
\item{rsids}{A vector of rsids}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
snpToInfo(rsids = "rs10")
snpToInfo(rsids = c("rs10","rs3","rs4","rs10047249"))
}
