% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_Fucntions.R
\name{SNP_To_Allele_Gtex}
\alias{SNP_To_Allele_Gtex}
\title{SNP_To_Allele_Gtex Function}
\usage{
SNP_To_Allele_Gtex(variant)
}
\arguments{
\item{rsids}{A vector of rsids}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
snpToAF(rsids = "rs10")
snpToAF(rsids = c("rs10","rs3","rs4","rs10047249"))
}
